feat: add limited state to generate-button
The Advice Slip API caches a random advice for 2 seconds, so repeated
calls within this time will return the same advice. In other words,
It's impossible to generate a new advice if one was already generated
2 seconds ago.

Therefore, it would be good to prevent users from generating an advice
and not receiving a new advice due to the API behavior.

This commit solves this issue by disabling generate-button
for 2 seconds after an advice is generated.